refactor: осуществляет декомпозицию главной страницы
