+++ src/pages/main/main.tsx
@@ -1,31 +1,15 @@
-import { useState } from 'react';
 import { useAppSelector } from '../../hooks';
-import { sortOffers } from '../../utils/utils';
-
 import Header from '../../components/header/header';
-import CardList from '../../components/card-list/card-list';
-import Map from '../../components/map/map';
 import CitiesList from '../../components/cities-list/cities-list';
-import Sorting from '../../components/sorting/sorting';
 import Loader from '../../components/loader/loader';
+import MainContainer from '../../components/main-container/main-container';
+import { getActiveCity, getOffers, getOffersLoadingStatus } from '../../store/selectors';
 
 function MainPage() {
-  const offers = useAppSelector((state) => state.offers);
-  const activeCity = useAppSelector((state) => state.city);
-  const currentSorting = useAppSelector((state) => state.sorting);
+  const offers = useAppSelector(getOffers);
+  const activeCity = useAppSelector(getActiveCity);
   const activeOffers = offers.filter((offer) => offer.city.name === activeCity.name);
-  const sortedOffers = sortOffers(activeOffers, currentSorting);
-  const isOffersLoading = useAppSelector((state) => state.isOffersLoading);
-
-  const [activeCardId, setActiveCardId] = useState<string | null>(null);
-
-  const handleCardMouseEnter = (id: string | null) => {
-    setActiveCardId(id);
-  };
-
-  const handleCardMouseLeave = () => {
-    setActiveCardId(null);
-  };
+  const isOffersLoading = useAppSelector(getOffersLoadingStatus);
 
   if (isOffersLoading) {
     return <Loader />;
@@ -40,17 +24,7 @@ function MainPage() {
           <CitiesList activeCityName={activeCity.name} />
         </div>
         <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{activeOffers.length} places to stay in {activeCity.name}</b>
-              <Sorting currentSorting={currentSorting}/>
-              <CardList offers={sortedOffers} onCardMouseEnter={handleCardMouseEnter} onCardMouseLeave={handleCardMouseLeave}/>
-            </section>
-            <div className="cities__right-section">
-              <Map city={activeCity} offers={activeOffers} activeCardId={activeCardId} mapPlace='cities'/>
-            </div>
-          </div>
+          <MainContainer offers={activeOffers} activeCity={activeCity}/>
         </div>
       </main>
     </div>
